backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
Pillow==10.1.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.25
stripe==7.8.0
twilio==8.10.0
sendgrid==6.10.0
reportlab==4.0.7
openpyxl==3.1.2
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
langchain==0.0.335
openai==1.3.5
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0